+++ NEWS.rst
@@ -5,6 +5,8 @@ Version History
 ------------------------
 
 * Added zstd.get_frame_parameters() to obtain info about a zstd frame.
+* Added ZstdDecompressor.decompress_content_dict_chain() for efficient
+  decompression of *content-only dictionary chains*.
 
 0.6.0 (released 2017-01-14)
 ---------------------------